Brings the issue-discovery harness (tools/issue-discovery/) plus the
storefront/buyer/token/release fixes onto dev.

Resolved storefront/Dockerfile: dev (67c4aec) and the feature branch
(2de2670) fixed the same uv find-links bug two ways. Kept dev's direct
"COPY .dist/ /.dist/" + "--find-links /.dist" approach and dropped the
feature branch's /dist + symlink workaround, which is now redundant.

+++ storefront/src/market_storefront/models/settle_models.py
@@ -48,7 +48,7 @@ class VerifyEscrowRequest(BaseModel):
     test getRecordFromChain in isolation before committing to settle.
     """
     seller_wallet: str = Field(description="Expected seller wallet address (recipient on-chain)")
-    agreed_price: float = Field(
+    agreed_price: int = Field(
         description=(
             "Expected absolute payment amount in base units of the payment "
             "token (the field name is retained from before the per-hour → "
